limiting fitting to 5000 junctions and vectorizing pval calculation

+++ majiq/src/io.pyx
@@ -244,7 +244,7 @@ cdef _get_extract_lsv_list(list list_of_lsv_id, list file_list, int msamples):
                     if lsv_type is None:
                         lsv_type = data['LSVs/%s' % lsv_id].attrs['type']
                         njunc = len(lsv_type.split('|')) -1
-                        lsv_cov = np.zeros(shape=(n_exp, njunc, msamples))
+                        lsv_cov = np.zeros(shape=(n_exp, njunc, msamples),  dtype=float)
 
                     assert data['LSVs/%s' % lsv_id].attrs['type'] == lsv_type, "ERROR lsv_type doesn't match " \
                                                                                "for %s" % lsv_id
